pyproject.toml
[tool.poetry]
name = "b2bpdsbackend"
version = "0.1.0"
description = "Bussines to Bussines miniErp PDS"
authors = ["Josue <urgosxd@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"
djangorestframework-simplejwt = "^5.3.1"
dj-rest-auth = "^5.0.2"
django-allauth = "^0.58.2"
dj-database-url = "^2.1.0"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"
whitenoise = {extras = ["brotli"], version = "^6.5.0"}
mysqlclient = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
